auth1stv

+++ service/index.js
@@ -1,15 +1,16 @@
 const Contact = require("./schemas/contact");
+const User = require("./schemas/user");
 
-const getAllContacts = async () => {
-  return Contact.find();
+const getAllContacts = async (userId) => {
+  return Contact.find({ owner: userId });
 };
 
-const getContactById = (id) => {
-  return Contact.findOne({ _id: id });
+const getContactById = (id, userId) => {
+  return Contact.findOne({ _id: id, owner: userId });
 };
 
-const createContact = ({ name, email, phone }) => {
-  return Contact.create({ name, email, phone });
+const createContact = ({ name, email, phone }, userId) => {
+  return Contact.create({ name, email, phone }, { owner: userId });
 };
 
 const updateContact = (id, fields) => {
@@ -20,8 +21,12 @@ const updateStatusContact = (id, body) => {
   return Contact.findByIdAndUpdate(id, body);
 };
 
-const removeContact = (id) => {
-  return Contact.findByIdAndRemove({ _id: id });
+const removeContact = (id, userId) => {
+  return Contact.findByIdAndRemove({ _id: id, owner: userId });
+};
+
+const createUser = ({ email, password }) => {
+  return User.create({ email, password });
 };
 
 module.exports = {
@@ -31,4 +36,5 @@ module.exports = {
   updateContact,
   removeContact,
   updateStatusContact,
+  createUser,
 };
